+++ lib/HttpServer/ApplicationEndpointServer.cpp
@@ -674,7 +674,7 @@ bool ApplicationEndpointServer::createSslContext () {
   Random::UniformCharacter r("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789");
   _rctx = r.random(SSL_MAX_SSL_SESSION_ID_LENGTH);
 
-  int res = SSL_CTX_set_session_id_context(_sslContext, (unsigned char const*) _rctx.c_str(), _rctx.size());
+  int res = SSL_CTX_set_session_id_context(_sslContext, (unsigned char const*) _rctx.c_str(), (int) _rctx.size());
 
   if (res != 1) {
     LOGGER_ERROR("ssl error: " << lastSSLError());
